+++ CHANGELOG.md
@@ -13,8 +13,9 @@
 
 ## Master
 
-- Append Swiftlint rule id to the Swiftlint danger messages [@f-meloni][]
-- Run Swiftlint from SPM by default when available [@f-meloni][]
+- Make swiftlint rule id code on the message by [@f-meloni][]
+- Append Swiftlint rule id to the Swiftlint danger messages by [@f-meloni][]
+- Run Swiftlint from SPM by default when available by [@f-meloni][]
 
 ## 1.0.0